обновление 2.0

+++ lazy_ilya/cities/views.py
@@ -3,22 +3,26 @@ import threading
 import traceback
 from typing import List, Dict, Any
 
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.contrib.auth.views import redirect_to_login
 from django.db.models import Q
-from django.http import HttpRequest, HttpResponse, JsonResponse, Http404
+from django.http import HttpRequest, HttpResponse, JsonResponse, Http404, FileResponse, HttpResponseNotFound, \
+    HttpResponseForbidden
 from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse_lazy
 from django.views import View
 from django.views.decorators.csrf import csrf_exempt
 
-from cities.forms import CityDataForm
-from cities.models import CityData, CounterCities
+from cities.forms import CityDataForm, CityInfoDoForm
+from cities.models import CityData, CounterCities, CityInfoDO
 from cities.utils.common_func.get_city_context import (
     get_all_cities,
     get_context_admin_cities,
 )
 from cities.utils.parser_word.globus_parser import GlobusParser
 from file_creator.utils.storage import OverwritingFileSystemStorage
+from lazy_ilya.settings import MEDIA_ROOT
 from lazy_ilya.utils.settings_for_app import logger, ProjectSettings
 
 
@@ -29,7 +33,7 @@ def base_view(request):
     return render(request=request, template_name="cities/admin-cities.html")
 
 
-class Cities(LoginRequiredMixin, View):
+class Cities(View):
     """
     Класс для обработки запросов, связанных с городами.
 
@@ -38,6 +42,16 @@ class Cities(LoginRequiredMixin, View):
     """
 
     login_url = reverse_lazy("myauth:login")
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.method in ["PUT", "DELETE"]:
+            if not request.user.is_authenticated:
+                # Для обычного запроса — редирект
+                if request.content_type == "text/html":
+                    return redirect_to_login(request.get_full_path(), self.login_url)
+                # Для JS/AJAX — 403 Forbidden
+                return JsonResponse({"status": "error", "message": "Authentication required"}, status=403)
+        return super().dispatch(request, *args, **kwargs)
 
     def get(self, request: HttpRequest) -> HttpResponse:
         """
@@ -56,44 +70,51 @@ class Cities(LoginRequiredMixin, View):
             context={**context},
         )
 
-    def put(self, request: HttpRequest, table_id: int, dock_num: int) -> JsonResponse:
+    def put(self, request: HttpRequest, table_id: int, dock_num: int | None = None) -> JsonResponse:
         """
         Обновляет информацию о городе.
 
         Args:
             request (HttpRequest): Объект запроса.
             table_id (int): ID таблицы.
-            dock_num (int): Номер доки.
+            dock_num (int|None): Номер доки| None.
 
         Returns:
             JsonResponse: Ответ с сообщением об успехе или ошибке.
         """
         try:
-            # Получаем город по ID таблицы и номеру доки
-            city = get_object_or_404(CityData, table_id=table_id, dock_num=dock_num)
-
             # Загружаем данные из тела запроса
             data = json.loads(request.body)
-            form = CityDataForm(data, instance=city)
-            if form.is_valid():
-                form.save()
-                # # Обновляем поля города
-                # city.location = data.get("location", city.location)
-                # city.name_organ = data.get("name_organ", city.name_organ)
-                # city.pseudonim = data.get("pseudonim", city.pseudonim)
-                # city.ip_address = data.get("ip_address", city.ip_address)
-                # city.work_time = data.get("work_time", city.work_time)
-                # city.some_number = data.get("some_number", city.some_number)
-                # city.save()
-                logger.bind(user=request.user.username).info(
-                    f"Произошло обновление города {city.name_organ} - {city.location}"
-                )
-                return JsonResponse({"status": "success"})
-            else:
-                logger.bind(user=request.user.username).error(
-                    f"Ошибка при создании города: {json.dumps(form.errors.get_json_data(), ensure_ascii=False, indent=2)}"
-                )
-                return JsonResponse({"errors": form.errors}, status=400)
+            if "pseudonim" in data:
+                # Получаем город по ID таблицы и номеру доки
+                city = get_object_or_404(CityData, table_id=table_id, dock_num=dock_num)
+                form = CityDataForm(data, instance=city)
+                if form.is_valid():
+                    form.save()
+                    logger.bind(user=request.user.username).info(
+                        f"Произошло обновление города {city.name_organ} - {city.location}"
+                    )
+                    return JsonResponse({"status": "success"})
+                else:
+                    logger.bind(user=request.user.username).error(
+                        f"Ошибка при создании города: {json.dumps(form.errors.get_json_data(), ensure_ascii=False, indent=2)}"
+                    )
+                    return JsonResponse({"errors": form.errors}, status=400)
+            elif "korr" in data:
+                city_do = get_object_or_404(CityInfoDO, pk=table_id)
+                form_do = CityInfoDoForm(data, instance=city_do)
+                if form_do.is_valid():
+                    form_do.save()
+                    logger.bind(user=request.user.username).info(
+                        f"Произошло обновление города {city_do.pk} - {city_do.korr}"
+                    )
+                    return JsonResponse({"status": "success"})
+                else:
+                    logger.bind(user=request.user.username).error(
+                        f"Ошибка при создании города: {json.dumps(form_do.errors.get_json_data(), ensure_ascii=False, indent=2)}"
+                    )
+                    return JsonResponse({"errors": form_do.errors}, status=400)
+
 
         except Http404:
             logger.bind(user=request.user.username).error(f"Город не найден")
@@ -114,7 +135,7 @@ class Cities(LoginRequiredMixin, View):
             return JsonResponse({"status": "error", "message": str(e)}, status=500)
 
     def delete(
-        self, request: HttpRequest, table_id: int, dock_num: int
+            self, request: HttpRequest, table_id: int, dock_num: int | None = None
     ) -> JsonResponse:
         """
         Удаляет город.
@@ -122,38 +143,54 @@ class Cities(LoginRequiredMixin, View):
         Args:
             request (HttpRequest): Объект запроса.
             table_id (int): ID таблицы.
-            dock_num (int): Номер доки.
+            dock_num (int|None): Номер доки. или Nonr
 
         Returns:
             JsonResponse: Ответ с сообщением об успехе или ошибке.
         """
         try:
-            # Получаем город по ID таблицы и номеру доки
-            city = get_object_or_404(CityData, table_id=table_id, dock_num=dock_num)
-            logger.bind(user=request.user.username).info(
-                f"Проиcходит удаление города {city.name_organ} - {city.location}"
-            )
-            if city:
+            if dock_num is not None:
                 try:
-                    counter_city = CounterCities.objects.get(
-                        dock_num=city
-                    )  # changed name to the name u have in related model
-                    counter_city.delete()  # Удаляем запись CounterCities
-                except CounterCities.DoesNotExist:
-                    # Если CounterCities не существует, ничего страшного, продолжаем
-                    pass
-                # Очищаем поля города
-                city.location = ""
-                city.name_organ = ""
-                city.pseudonim = ""
-                city.letters = False
-                city.writing = False
-                city.ip_address = ""
-                city.some_number = ""
-                city.work_timme = ""
-                city.save()
+                    # Получаем город по ID таблицы и номеру доки
+                    city = CityData.objects.get(table_id=table_id, dock_num=dock_num)
+                    logger.bind(user=request.user.username).info(
+                        f"Проиcходит удаление города {city.name_organ} - {city.location}"
+                    )
+                    if city:
+                        try:
+                            counter_city = CounterCities.objects.get(
+                                dock_num=city
+                            )  # changed name to the name u have in related model
+                            counter_city.delete()  # Удаляем запись CounterCities
+                        except CounterCities.DoesNotExist:
+                            # Если CounterCities не существует, ничего страшного, продолжаем
+                            pass
+                        # Очищаем поля города
+                        city.location = ""
+                        city.name_organ = ""
+                        city.pseudonim = ""
+                        city.letters = False
+                        city.writing = False
+                        city.ip_address = ""
+                        city.some_number = ""
+                        city.work_timme = ""
+                        city.save()
 
-            return JsonResponse({"status": "success"})
+                    return JsonResponse({"status": "success"})
+                except CityData.DoesNotExist:
+                    pass
+            # Пытаемся удалить CityInfoDO, если dock_num не передан или CityData не найден
+            try:
+                city_do = CityInfoDO.objects.get(pk=table_id)
+                logger.bind(user=request.user.username).info(
+                    f"Удаление CityInfoDO: ID {city_do.pk}, korr: {city_do.korr}"
+                )
+                city_do.delete()
+                return JsonResponse({"status": "success"})
+
+            except CityInfoDO.DoesNotExist:
+                # Ни CityData, ни CityInfoDO не найдены
+                raise Http404("Город не найден")
 
         except Http404:
             logger.bind(user=request.user.username).error(f"Город не найден")
@@ -179,8 +216,8 @@ class CitiesAdmin(LoginRequiredMixin, UserPassesTestMixin, View):
     def test_func(self):
         # Проверяем, что пользователь в группе admin
         return (
-            self.request.user.groups.filter(name="admin").exists()
-            or self.request.user.is_superuser
+                self.request.user.groups.filter(name="admins").exists()
+                or self.request.user.is_superuser
         )
 
     def handle_no_permission(self):
@@ -269,8 +306,8 @@ class CityInfoView(LoginRequiredMixin, UserPassesTestMixin, View):
     def test_func(self):
         # Проверяем, что пользователь в группе admin
         return (
-            self.request.user.groups.filter(name="admin").exists()
-            or self.request.user.is_superuser
+                self.request.user.groups.filter(name="admins").exists()
+                or self.request.user.is_superuser
         )
 
     def handle_no_permission(self):
@@ -330,18 +367,33 @@ class CityInfoView(LoginRequiredMixin, UserPassesTestMixin, View):
         :return: JSON-ответ с результатом создания или ошибками валидации
         """
         data = json.loads(request.body)
-        form = CityDataForm(data)
-        if form.is_valid():
-            obj: CityData = form.save()
-            logger.bind(user=request.user.username).info(
-                f"Город успешно создан: {obj.id}"
-            )
-            return JsonResponse({"created": True, "id": obj.id})
-        else:
-            logger.bind(user=request.user.username).error(
-                f"Ошибка при создании города: {json.dumps(form.errors.get_json_data(), ensure_ascii=False, indent=2)}"
-            )
-            return JsonResponse({"errors": form.errors}, status=400)
+        if "pseudonim" in data:
+            form = CityDataForm(data)
+            if form.is_valid():
+                obj: CityData = form.save()
+                logger.bind(user=request.user.username).info(
+                    f"Город успешно создан: {obj.id}"
+                )
+                return JsonResponse({"created": True, "id": obj.id})
+            else:
+                logger.bind(user=request.user.username).error(
+                    f"Ошибка при создании города: {json.dumps(form.errors.get_json_data(), ensure_ascii=False, indent=2)}"
+                )
+                return JsonResponse({"errors": form.errors}, status=400)
+
+        elif "korr" in data:
+            form = CityInfoDoForm(data)
+            if form.is_valid():
+                obj = form.save()
+                logger.bind(user=request.user.username).info(
+                    f"Город успешно создан: {obj.id}"
+                )
+                return JsonResponse({"created": True, "id": obj.id})
+            else:
+                logger.bind(user=request.user.username).error(
+                    f"Ошибка при создании города: {json.dumps(form.errors.get_json_data(), ensure_ascii=False, indent=2)}"
+                )
+                return JsonResponse({"errors": form.errors}, status=400)
 
     def put(self, request: HttpRequest) -> JsonResponse:
         """
@@ -427,3 +479,87 @@ def increment_city_counters(request: HttpRequest) -> JsonResponse:
             f"Ошибка при обновлении счетчиков: {str(e)}"
         )
         return JsonResponse({"error": str(e)}, status=400)
+
+
+class CityDownload(LoginRequiredMixin, UserPassesTestMixin, View):
+    """
+    View для запуска фонового создания файла с данными города.
+
+    Доступ разрешён только администраторам и суперпользователям.
+    """
+
+    login_url = reverse_lazy("myauth:login")
+
+    def test_func(self) -> bool:
+        """
+        Проверяет, входит ли пользователь в группу 'admins' или является суперпользователем.
+
+        Returns:
+            bool: True, если пользователь администратор или суперпользователь, иначе False.
+        """
+        return (
+                self.request.user.groups.filter(name="admins").exists()
+                or self.request.user.is_superuser
+        )
+
+    def handle_no_permission(self) -> HttpResponse:
+        """
+        Обрабатывает отказ в доступе. Если пользователь не авторизован, редиректит на логин.
+        Если авторизован, но не имеет прав — возвращает 403 Forbidden.
+
+        Returns:
+            HttpResponse: Редирект или сообщение об ошибке 403.
+        """
+        if not self.request.user.is_authenticated:
+            return redirect(self.login_url)
+
+        from django.http import HttpResponseForbidden
+        return HttpResponseForbidden(
+            "Доступ запрещён, только админ может сюда заходить"
+        )
+
+    def get(self, request: HttpRequest) -> JsonResponse:
+        """
+        Обрабатывает GET-запрос, запускает создание файла в отдельном потоке.
+
+        Args:
+            request (HttpRequest): Объект запроса от клиента.
+
+        Returns:
+            JsonResponse: Ответ об успешном запуске задачи.
+        """
+        logger.bind(user=request.user.username).info(
+            "Запуск создания файла в отдельном потоке"
+        )
+        threading.Thread(target=GlobusParser.create_globus).start()
+        logger.bind(user=request.user.username).info(
+            "Файл начал создание успешно"
+        )
+        return JsonResponse({"status": "success"}, status=200)
+
+
+@login_required
+def download_file(request: HttpRequest, filename: str) -> HttpResponseForbidden | FileResponse | HttpResponseNotFound:
+    """
+    Позволяет администратору или суперпользователю скачать файл по имени.
+
+    Args:
+        request (HttpRequest): Объект HTTP-запроса.
+        filename (str): Имя файла для скачивания.
+
+    Returns:
+        HttpResponse: Ответ с файлом для загрузки или сообщение об ошибке:
+            - 403 Forbidden, если у пользователя нет прав.
+            - 404 Not Found, если файл не найден.
+    """
+    user = request.user
+    if not (user.is_superuser or user.groups.filter(name='admins').exists()):
+        return HttpResponseForbidden("Доступ запрещён")
+
+    file_path = MEDIA_ROOT / filename
+    if file_path.exists() and file_path.is_file():
+        response = FileResponse(open(file_path, 'rb'))
+        response['Content-Disposition'] = f'attachment; filename="{filename}"'
+        return response
+    else:
+        return HttpResponseNotFound("Файл не найден")
